% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{islamic_pal}
\alias{islamic_pal}
\title{Islamic palette with ramped colors}
\usage{
islamic_pal(palette = "samarqand", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from \code{\link{islamic_palettes}} list.}

\item{alpha}{Transparency.}

\item{reverse}{If TRUE, the direction of the colors is reversed.}
}
\description{
Islamic palette with ramped colors
}
\examples{
library(scales)
show_col(islamic_pal()(10))

filled.contour(volcano,color.palette = islamic_pal(), asp=1)
}
