% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{islamic_palettes}
\alias{islamic_palettes}
\title{A list of color palettes inspired by Islamic art}
\format{An object of class \code{list} of length 16.}
\usage{
islamic_palettes
}
\description{
This is a named list of character vectors with hex colors hand picked from
photos Islamic art to give some feel of the original art. Most of them are
from architecture. Art from the following cities were selected here:
Abu Dhabi, Alhambra, Cordoba, Damascus, Fes, Istanbul, Jerusalem, Konya,
Samarqand, and Shiraz. There is also a palette for an Ottoman tile color
scheme
}
\details{
Efforts were made to avoid overly similar colors in the same palettes as well
as colors that are too light. Some cities were not included not because they
are not awesome, but because it's difficult to make a good palette from
landmarks of those cities.

These are the names of palettes available and where they come fome:

\describe{
\item{shiraz}{Nasir al-Mulk Mosque, Shiraz, Iran}
\item{shiraz2}{The Jerrāḥ Pasha Quran Quran, Shiraz}
\item{samarqand}{Samarqand Shah-i-Zinda Necropolis}
\item{samarqand2}{Samarqand Bibi-Khanym Mosque tile}
\item{abu_dhabi}{Abu Dhabi Sheikh Zayed Grand Mosque}
\item{istanbul}{Istanbul Sultan Ahmed Mosque (Blue Mosque) exterior}
\item{istanbul2}{Blue Mosque interior}
\item{istanbul3}{Hagia Sophia interior}
\item{konya}{Konya Mevlana Museum}
\item{jerusalem}{Dome of the Rock}
\item{fes}{Panorama of Fes, Morocco}
\item{fes2}{Qarawiyyin University, Fes}
\item{alhambra}{Geometric tile of al-Andalus style}
\item{cordoba}{Great Mosque of Cordoba}
\item{damascus}{Umayyad Mosque, Damascus, Syria}
\item{ottoman}{Ottoman tile}
}
}
\keyword{datasets}
