% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_islamic}
\alias{scale_color_islamic}
\alias{scale_colour_islamic}
\title{Setup color palette for ggplot2}
\usage{
scale_color_islamic(..., palette = "samarqand", discrete = TRUE,
  alpha = 1, reverse = FALSE)

scale_colour_islamic(..., palette = "samarqand", discrete = TRUE,
  alpha = 1, reverse = FALSE)
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[ggplot2]{scale_color_gradientn}}.}

\item{palette}{Choose from \code{\link{islamic_palettes}} list.}

\item{discrete}{Whether to use a discrete color palette.}

\item{alpha}{Transparency.}

\item{reverse}{Logical, whether to reverse the order of the colors.}
}
\description{
Setup color palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl))) +
  scale_colour_islamic(palette="ottoman")
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hp)) +
  scale_colour_islamic(palette="shiraz", discrete = FALSE)
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_colour_islamic(palette="jerusalem")
ggplot(diamonds) + geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_islamic()
}
