% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_islamic}
\alias{scale_fill_islamic}
\title{Setup fill palette for ggplot2}
\usage{
scale_fill_islamic(..., palette = "samarqand", discrete = TRUE,
  alpha = 1, reverse = TRUE)
}
\arguments{
\item{...}{additional arguments to pass to scale_color_gradientn}

\item{palette}{Choose from 'islamic_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{pass through parameter to \code{viridis}}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
Setup fill palette for ggplot2
}
