% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{islamic_pal}
\alias{islamic_pal}
\title{islamic palette with ramped colours}
\usage{
islamic_pal(palette = "samarqand", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'islamic_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
islamic palette with ramped colours
}
\examples{
library(scales)
show_col(islamic_pal()(10))

filled.contour(volcano,color.palette = islamic_pal(), asp=1)
}
