% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_islamic}
\alias{scale_fill_islamic}
\title{Setup fill palette for ggplot2}
\usage{
scale_fill_islamic(..., palette = "samarqand", discrete = TRUE,
  alpha = 1, reverse = TRUE)
}
\arguments{
\item{...}{Additional arguments to pass to \code{\link[ggplot2]{scale_fill_gradientn}}.}

\item{palette}{Choose from \code{\link{islamic_palettes}} list.}

\item{discrete}{Whether to use a discrete color palette.}

\item{alpha}{pass through parameter to \code{viridis}}

\item{reverse}{If TRUE, the direction of the colors is reversed.}
}
\description{
Setup fill palette for ggplot2
}
